Rename things to make auto-complete happier
